SDL2 stuff use double rather than float

+++ fenton.h
@@ -7,7 +7,7 @@ F_Quit (void);
 extern void
 F_Error (const char *fmt, ...);
 
-extern float frametime;
+extern double frametime;
 extern unsigned int elapsedtime_ms;
 
 #endif /* __FENTON_H__ */
